+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - New `retry` processor.
 - New `noop` cache.
 - Field `targets_input` added to the `azure_blob_storage` input.
+- New `reject_errored` output.
 
 ### Fixed